+++ packages/editor/CHANGELOG.md
@@ -18,6 +18,7 @@
 - Automated test to check `font color` WP data compatibility if variables not found. 
 - Automated test to check `background color` WP data compatibility if variables not found. 
 - Automated test to check `border color` WP data compatibility if variables not found. 
+- Automated test to check `font size` WP data compatibility if variables not found. 
 
 
 ## 2.0.1 (2025-07-22)